[open][neva_upstreamable][watchdog] Watchdog writes callstack of stucked thread at PmLog
:Release Notes:
Watchdog writes the callstack of stucked thread at PmLog

:Detailed Notes:
When Watchdog kills process due to stuck, the callstack is retrieved
at crash report.
However, sometimes this callstack is not for the actual stucked thread.
To remain the callstack of the actual stucked thread, watchdog sends
kernal signal to stucked thread and retrieve the callstack by itself
and writes it to PmLog.

:Testing Performed:
1. generate sutck
2. wait until watchdog recovers the stucked process
3. check there is a callstack of stucked thread in PmLog

:QA Notes:
Test-cases: GT-27299

:Issues Addressed:
[PLAT-80051] Port "[open][neva_upstreamable][watchdog] Watchdog writes callstack of stucked thread at PmLog"

Change-Id: I9c7caea3256f0e194183e4e6870dbccad604b3b2
Originally-Reviewed-on: http://gpro.lgsvl.com/235781
Reviewed-on: http://gpro.lge.com/243560