[open][neva_upstream][memory][gpu] Flush shared image interface when ReclaimResources
:Release Notes:
Flush shared image interface when ReclaimResources

:Detailed Notes:
When ReclaimResources, SharedImageInterfaceProxy::DestroySharedImage
calls EnqueueDeferredMessage so it doesn't trigger releasing gpu
resources immediately. When it is not visible we aggressively flush to
release shared image in GPU process.
This patch also contains a part of upstream patch of
https://chromium-review.googlesource.com/c/chromium/src/+/1616526
to add SharedImageInterface::Flush method.

:Test Scenario:
1. Enable --enable-aggressive-release-policy
2. Launch com.webos.app.enactsampler application.
3. Launch com.webos.app.userguide application.
4. Collect chrome tracing with memory-infra during about 1 second.
5. All resources in CC in com.webos.app.enactsampler process should be
cleared(zero) and gpu category of Browser process should be less than 5mb.
(Previously it was about 20mb)

:QA Notes:
Test-cases: GT-26790

:Issues Addressed:
[NEVA-4000] Neva stable 1.1 consumes more GPU memory than RP chromium68

Change-Id: I459b5089bcf44231859ac6c67d545b26b0d05ee0
Originally-Reviewed-on: http://gpro.lge.com/249480
Reviewed-on: http://gpro.lge.com/249818
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>