[open][neva_upstreamable][ozone] Fix transition from minimized to fullscreen
:Release Notes:
Fix transition from minimized to fullscreen.

:Detailed Notes:
When we use deep linking to open an app, and then we
close the app using the close button, the
app is minimized (sent to background). That makes the app
minimized.

When opening again app (through deep linking or just opening
the app in the HOME ribbon), the transition that happens is
directly from minimized to fullscreen state. This patch adds
that case, reenabling rendering and activating several parts of the
stack.

:Testing Performed:
Case 1:
* Open LG Store, press search and search "JPG"
* Select a JPG image in the list
* Close the photo using bottom bar close button
* Launch photo&video

Case 2:
* Open LG Store, press search and search "JPG"
* Select a JPG image in the list
* Close the photo using bottom bar close button
* Go to LG Store, search "JPG"
* Open a JPG image in the list

In both cases it was freezing with the previously open photo. With the
fix both cases work as expected.

:QA Notes:
Test-cases: GT-27717, GT-27718

:Issues Addressed:
[PLAT-80946] Photo&Video app is displayed with black screen when the app is relaunched.

Change-Id: I9dc7ebea7ce53863dc11cef4aba884fd1d4f5be8
Originally-Reviewed-on: http://gpro.lge.com/249482
Reviewed-on: http://gpro.lge.com/249817
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>